pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oak-vision-system"
version = "0.1.0"
description = "OAK视觉检测+运动控制的机械臂抓取系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "depthai==2.30.0.0",
    "opencv-python==4.8.1.78",
    "numpy==1.24.3",
    "scipy==1.11.4",
    "python-can==4.2.2",
    "structlog==23.2.0",
    "asyncio-mqtt==0.16.1",
]

[project.optional-dependencies]
yaml = [
    "ruamel.yaml>=0.17.0",  # 推荐：支持注释保持和格式保持
]
yaml-basic = [
    "pyyaml>=6.0",  # 基础 YAML 支持（不保留注释）
]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
    "hypothesis==6.92.0",
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["oak_vision_system*"]

[tool.pytest.ini_options]
testpaths = ["oak_vision_system/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]